reorganize code, minor cleanup
