+++ src/lambdaActor/astToMetadata.js
